BUG: Fix main window toolbar buttons style to support Qt::ToolButtonTextUnderIcon
The tool button style is controlled in the Advanced Settings General panel.

From Qt 4.6.3 doc:
QToolBar::toolButtonStyle - "Note that if you add a QToolButton with the addWidget() method, it will not get this button style"

The following behavior does not seem  to work:
QToolButton::toolButtonStyle - QToolButton automatically connects this slot to the relevant signal in the QMainWindow in which is resides.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@18970 3bd1e089-480b-0410-8dfb-8563597acbee