BUG: crash in vtkImageResliceMask when pipeline stops using backgroundMask
Grayscale images use backgroundMask (vtkImageResliceMask 2nd output), but
DTIs don't need backgroundMask. In lightbox mode, the update_extent of
images is important: (0,w,0,h,0,0) for the 1st image, (0,w,0,h,1,1) for
the 2nd image (0,w,0,h,2,2) for the 3rd and so on.
When vtkImageResliceMask populates the output of the 2nd image, it
needs to access the scalar pointer of the output image and background mask.
However, if the request_update_extent of the background image is out of date
(because not used in the pipeline), it fails to retrieve the pointer and
leads to a crash later (should probably be fixed eventually).
Closes #1689

git-svn-id: http://svn.slicer.org/Slicer4/trunk@19257 3bd1e089-480b-0410-8dfb-8563597acbee