BUG: Loading a scene which had Slices linked did not restore properly.
During a restore, the SliceLinkLogic would try to propagate the attributes
to all the Slice ad SliceComposite nodes.

Ties into the Scene events to manage/disable SliceLinkLogic during loading
a scene or reverting a SceneView.

Changed the Broadcasting ivar from a flag to a counter so we can track the
nesting of StartBatchProcessing, StartImport, ...

Added priorities to events in ObserverManager and AbstractLogic so that the
SliceLinkLogic is always the LAST thing told about the EndBatchProcessing,
EndImport, EndRestore events so that only then will it enable the
SliceLinkLogic.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@19319 3bd1e089-480b-0410-8dfb-8563597acbee