ENH: Illustrate how vtkMRMLNode::CreateNodeByClass and SetReferenceCount should be used from python
This is a addendum to the mail [1] previously send on the list.

The reference count shouldn't be decreased using the naive approach where
it is set to one. Indeed, internally, the reference count could be any
number greater than one. It doesn't have to be 2.

Considering this last remark, using the following approach is *REQUIRED*,
otherwise it will lead to a CRASH of the application.

 n = slicer.mrmlScene.CreateNodeByClass('vtkMRMLViewNode')
 slicer.mrmlScene.addNode(n)
 n.SetReferenceCount(n.GetReferenceCount() - 1)

Even better, this specific example can be simplified using the following
syntax, this is the *RECOMMENDED* approach, it will prevent bug, memory
leaks and crashes:

  n = slicer.mrmlScene.addNode(slicer.vtkMRMLViewNode())

And if the name of the node is generated at runtime, this could be done:

  n = eval('slicer.mrmlScene.AddNode(slicer.%s())' % 'vtkMRMLViewNode')


[1] http://massmail.spl.harvard.edu/public-archives/slicer-devel/2011/007513.html

git-svn-id: http://svn.slicer.org/Slicer4/trunk@19772 3bd1e089-480b-0410-8dfb-8563597acbee