ENH: Simplify extension CMakeLists.txt moving description file generation.
The call to "slicerFunctionGenerateExtensionDescription" now happens
into SlicerExtensionCPack file. It has the inconvenient of
obfuscating where the generation of the extension description file
occurs but I believe it un-clutter the CMakeLists of each extension file.

This change is somewhat backward compatible. Indeed, if extension
maintainer does not update its CMakeLists, the file will be generated
two times.

It's possible to skip the generation of the file by setting the variable
Slicer_CPACK_SKIP_GENERATE_EXTENSION_DESCRIPTION to TRUE.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@19815 3bd1e089-480b-0410-8dfb-8563597acbee