COMP: Ensure environment is set when running "out-of-source" modules tests.
In the case of extensions bundling modules or module simply built against
a Slicer build-tree, it's common to have either an executable (CLI)
or a plugin (loadable module) that depends on other libraries provided
by the module itself (logic and MRML libraries for example).

In order to successfully be able to run the tests, a file containing
additional settings specific to the module is configured. These settings
include:
  Paths
  LibraryPaths
  EnvironmentVariables (for example PYTHONPATH)

These settings are similar to the values used to update SlicerLauncherSettings.ini
when an extension is installed using the  extension manager model.

Since the LAUNCH_COMMAND is updated with the option '--launcher-additional-settings'
associated with the additional settings file, the tested executables
and libraries should be able to resolve their dependencies properly.

Note also that PYTHONPATH is updated so that python module associated with
a module could be imported. For example, the scripted module.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@19827 3bd1e089-480b-0410-8dfb-8563597acbee