BUG: vtkMRMLDisplayableNode intermediate state was caught by Model displayable manager
When a null node ID was set to the displayable node, the DisplayNodeIDs had its element erased, and while unobserving the display node (vtkMRMLDisplayableNode::SetAndObserveNthDisplayNode(n,0)), DisplayModifiedEvent was fired and caught by the model displayable manager before the DisplayNodes element was erased from the list.
In the displayable manager, GetDisplayNodes() was called which fired an assert as the lists DisplayNodeIDs and DisplayNodes had a different size.
For the sake of testing, GetInternalDisplayNodes is exposed to check the internal values of the displayable node.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@19984 3bd1e089-480b-0410-8dfb-8563597acbee