ENH: Speedup scene view store
In vtkMRMLSceneViewNode::StoreScene(), by adding nodes directly into the scene collection, the vtkMRMLScene::NodeIDs cache was out of date.
This has the drawback to recompute the entire cache next time GetNodeByID is called. And GetNodeByID happens every time vtkMRMLDisplayableNode::CopyWithoutModifiedEvent() is called, the larger the scene is, the slower it is.
Instead, the node is added properly into the scene which updates the cache by just adding 1 entry for each added node. No need to populate the cache for each GetNodeByID.
The speed-up is quite significant (on MSVC 2008-debug): for 5000 display/displayable pairs it goes from 206s to 23s. For 100 pairs: 1.5s -> 0.4s
This speed-up also remove the warnings (vtkMRMLScene.cxx:2999 -> "There is danger here") generated by the tests in vtkMRMLSceneViewNodeStoreSceneTest

git-svn-id: http://svn.slicer.org/Slicer4/trunk@20068 3bd1e089-480b-0410-8dfb-8563597acbee