COMP: Do not compare CMAKE_*_COMPILER when building extension using MSVC
It resumes that on windows, CMAKE_{C,CXX}_COMPILER is set to "cl" and not
the path to the associated executable.

Waiting we find a solution, in order to re-enable the successful build
of experimental, continuous and nightly extension, the comparison is
skipped for MSVC. Issue #2304 has been created to capture that.
(See http://www.na-mic.org/Bug/view.php?id=2304)


See http://www.na-mic.org/Bug/view.php?id=2279

Fixes #2279

git-svn-id: http://svn.slicer.org/Slicer4/trunk@20581 3bd1e089-480b-0410-8dfb-8563597acbee