ENH: implement a confidence option for dicom loadables
Since multiple plugins may be able to load a give dicom series,
plugins now have the option of specifying their confidence that
they are the 'right' plugin.  For example, an RTDOSE series may
be readable as either a scalar volume or an RTDOSE object.  If the
modality is RTDOSE, then that plugin can set the confidence to 1.0,
where the scalar volume will have the default 0.5.  Only the plugins
with the hightest confidence values will be selected by default
for the user.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@20653 3bd1e089-480b-0410-8dfb-8563597acbee