BUG: database attribute did not need to be a QSharedPointer
As a shared pointer things worked until we try to change the
database directory with the directory button.  When doing this,
we get a message like the following and then a crash.
QSharedPointer: pointer 0x2384d70 already has reference counting

Making this a regular pointer seems to work, since the memory is
handled by the python code in the DICOM module.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@20873 3bd1e089-480b-0410-8dfb-8563597acbee