ENH: implement feature for request #1876 - DICOM notification
Now series are shown in the dicom module in order of their
insertion date, so if you are receiving data periodically from
a scanner the most recent studies will be listed first.

Clicking on a series will give the option to load it.

Also a status message is displayed in the application main
window indicating when the most recent database insertion
happened.  This will be visible when new data is received, for
example via a network transaction, even when you are working
in a different module.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@20908 3bd1e089-480b-0410-8dfb-8563597acbee