BUG: Update midas_api_escape_for_url to urlencode all recommended characters
This is commit improve the url encoding of parameter for sake of completeness
only. It doesn't fix the problem.

It resumes that URL passed as parameter to the midas api can't have any
ampersand. Indeed, the value following the ampersand will be considered
as a parameter of the API call itself.

See #2643 / http://www.na-mic.org/Bug/view.php?id=2643

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21199 3bd1e089-480b-0410-8dfb-8563597acbee