BUG: #1935, #1931, #1926, #1780. Interpolation controls enabled.
r19320 blocks signals from the volume selection combo boxes to
avoid them broadcasting volume selections when viewers are linked.
r19320 solved an issue where changing to the View Controllers module
created new Slice Controllers which when configured would broadcast
to the other Slice Views. (I believe this is an issue with how the
node combo boxes emit signals).

A side effect of r19320 was the the controls for volume interpolation
were not being activated since the signals were not being sent.

Added a few explicit calls to make sure the buttons are enabled appropriately.

Still appears as though the intial state of the buttons is not being
configured properly, particularly as the state in the display node
is changed through other means.

Observation:

Interpolation is a display node properly, so modifying the interpolation
on a volume shown in the foreground, modifies the interpolation of the volume
in all Slice Viewers, regardless of the layer.

If linking is on, then this behavior occurs PLUS the interpolation is modified
for all volumes at the SAME LAYER.  This in turn can trigger the first
behavior.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21247 3bd1e089-480b-0410-8dfb-8563597acbee