BUG: stop the annotation tree view from auto expanding lists
Don't use calling setMRMLScene on the tree view to refresh the tree, 
as it resets the expand level on the tree view, only call it when 
the scene is different.
This stopped the auto added top level all annotations list and the fiducials/ruler
list nodes from expanding, so invoke an event from the logic when those default
hierarchy nodes are added to trigger the tree to expand them (using the
Expanded value on the hierarchy node).
Also expand the hierarchy node when it's added by the user via the add
hierarchy button.

Fixes issue #1735


git-svn-id: http://svn.slicer.org/Slicer4/trunk@21531 3bd1e089-480b-0410-8dfb-8563597acbee