BUG: handle fact that tensors input can come with either 6 or 9 components
Internally slicer (vtk) always represents tensors as 9 component, but
that conversion happens after the data is read in.  So this optimization
needs to check for both cases.

Note that the py_SlicerMRBTest failed as a result of this optimization,
but nobody checked (it happened during the SLC project week, so
things were busy and other things broke as well).

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21627 3bd1e089-480b-0410-8dfb-8563597acbee