ENH: update dataprobe even when mousing over annotation
Increase the priority to get all events, even if they will later
be used by another observer.  Also be sure to observe the interactor
rather than the interactor style so the priority will work with
respect to widgets which observe the interactor itself.  The
interactor style is observing the interactor so this is the same
for the purposes of the data probe.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21741 3bd1e089-480b-0410-8dfb-8563597acbee