ENH: Add WandEffect to segmentation editor
This tool was a demonstration for creating an editor extension, but
some users found it to be a valuable addition, so Andras suggested
adding it to the core of slicer.  This version exposes additional options,
such as segmenting in 3D and also using a fixed threshold for better
precision when segmenting.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21823 3bd1e089-480b-0410-8dfb-8563597acbee