ENH: Updates for BRAINSTools and ITKv4
This is a set of updates to try to address reported
issues with 32bit windows builds related to the new
double-convert to ASCII functionality.

Worked with ITKv4 team (Bill Lorensen and Kent Williams mostly) to
add the DoubleConversion featues directly to ITKv4.  This simplyfies
adding this functionality to other tools (like BRAINSTools, Slicer,
ANTs, etc..), and also provides more extensive testing of ITKv4.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21839 3bd1e089-480b-0410-8dfb-8563597acbee