BUG: External link in extension manager now open in browser. Fixes #3045
To determine if a link should be open in the extension manager or in an
external browser, the "host" of the clicked url and the server url are
compared.

See http://qt-project.org/doc/qt-4.8/qurl.html#setAuthority for more details.

See http://www.na-mic.org/Bug/view.php?id=3045

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21897 3bd1e089-480b-0410-8dfb-8563597acbee