BUG: Observe scene after being populated
If a node is added into the scene while it is being populated
(updateScene() calls populateScene()), the added node might be added in
the wrong location and it might trigger modified event which confuse
the model.
Such added node can happen if a qMRMLNodeComboBox does not accept "None",
so it selects the first valid node as soon as it is added into the model
(by populateScene), selecting that node could trigger a node addition into
the scene (if someone observe the currentNodeChanged signal of the combobox).

git-svn-id: http://svn.slicer.org/Slicer4/trunk@21922 3bd1e089-480b-0410-8dfb-8563597acbee