COMP: Allow multiple configure with "ConfigurePrerequisites" component
Following commit r22038 [1], the use of CMAKE_OSX_* variables set
by Slicer has been simplified with the introduction of the
"ConfigurePrerequisites" components.

It turns out that the associated CMake logic was flawed and wasn't
working as expected when the project was configured twice.

This commit ensures the "Slicer_USE_*" variables associated with the
components are reset before checking the variable "Slicer_FIND_COMPONENTS".

It also ensures that the execution return if component "ConfigurePrerequisites"
is defined and "Slicer_PREREQUISITES_CONFIGURED" is already set.

Co-authored-by: Marine Clogenson <marine.clogenson@epfl.ch>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22115 3bd1e089-480b-0410-8dfb-8563597acbee