COMP: Make sure extension fixup considers extension "support libraries"
Before this commit, libraries that were not (python, cli or loadable module)
 related were systematically fixed with "@fixup_path@/@Slicer_LIB_DIR@".

This commit changes this so that "support libraries" provided by
an extension are properly fixed up.

Note that the approach considered here will have to be revisited to ensure
that extension depends on support libraries provided by other extension
are properly fixed up.

Co-authored-by: Francois Budin <fbudin@unc.edu>
Signed-off-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22116 3bd1e089-480b-0410-8dfb-8563597acbee