BUG: Turn off projections when annotations are not visible
Projections of annotations were not "hidden" with annotation.

This fix ensure the widget representing an annotation's projection
is "hidden" when an annotation visibility is turned off.

This commit is a follow-up of r22176 ensuring projections are deleted
when an annotation is deleted.

Fixes #3214

From: Laurent Chauvin <lchauvin@bwh.harvard.edu>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22182 3bd1e089-480b-0410-8dfb-8563597acbee