ENH: Add support for node reference templates
In vtkMRMLSelectionNode, the exact name of the reference roles of a unit is
not known until the unit is set and observed. (it's a dynamic number of
unit reference roles). By using a special key '/' in the reference role
(e.g. "unit/"), it is possible to have multiple reference roles
(e.g. "unit/length", "unit/time"...).
Issue #2727

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22188 3bd1e089-480b-0410-8dfb-8563597acbee