ENH: Improve qMRMLUnitWidget interface
QFlags were added to this widget so one can choose what properties are
visible or not. This is what is done in the qSlicerUnitsSettingsPanel to
hide the advanced properties of the unit to the user. By default, the user
can only modify the precision property.
If he choose to show more properties, a warning stating that only display
values are changed when modifying an unit is shown along with the other
properties.

Style wise, the groupbox title now has its first letter in uppercase to
follow the GUI conventions and a horizontale line in the
qMRMLUnitNodeWidget between the precision and the rest of the properties.

See Issue #3137

From: Johan Andruejohl <johan.andruejohl@kitware.com>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22190 3bd1e089-480b-0410-8dfb-8563597acbee