ENH: Group appearance related settings together
The settings in General that are related to the slicer's look were moved
to the style setting panel. At this occasion, this panel was renamed
Appearance.
This allow a better grouping of the application customization option for
an easier user experience.

Also used the occasion to add tooltips to these options.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22236 3bd1e089-480b-0410-8dfb-8563597acbee