BUG: #2326 mark a node as modified if the name has changed
When the user has changed the name of a node this is
reflected in the suggested name of the file (see issue #3125).
With this fix the node is also marked as modified so that
it will be checked for saving.

This means that changing the name of a node in the scene will
cause it to be saved to a new file on disk which is a slight change
in behavior and may cause people to use more disk space.  But in
general I think this is safer and more clear, since file names will
always match the name of the node in the scene and users will
prefer that I believe.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22260 3bd1e089-480b-0410-8dfb-8563597acbee