BUG: Set module name / objectName in GenericTests
In addition to be consistent with what is done in the "instantiateModule"
method of the "qSlicerAbstractModuleFactoryManager".

It also ensure that the "ModuleShareDirectory" set in the "logic" method
of the "qSlicerAbstractCoreModule" class is done using the expected
value for the module name.

This will allow generic tests instantiating module using
"GetModuleShareDirectory" method to get the correct value (i.e Reporting
module)

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22429 3bd1e089-480b-0410-8dfb-8563597acbee