BUG: fix typo in AdditionalInitializeStep method name
Fixing a mis spelling in the additional initialize step method name.
In vtkMRMLAbstractDisplayableManager it's defined as:
AdditionalInitializeStep
but vtkMRMLCameraDisplayableManager was using:
AdditionnalInitializeStep
with an extra n. Using that signature, it was never getting called
so when new layout windows were being created, the active camera
was not getting set on the renderer.

Issue #1690



git-svn-id: http://svn.slicer.org/Slicer4/trunk@22611 3bd1e089-480b-0410-8dfb-8563597acbee