BUG: refactor code to remove unnecessary error message
There are two possible ways to obtain the results from PETStandardUptakeValueComputation,
having them written out to a .csv file or to a string parameter file that gets
parsed and loaded into the Slicer GUI. Only print out an error and fail if
neither parameter is specified. 
Checks for empty strings and doesn't try to open the files, so the unable
to open error messages will only occur if a non empty file name fails
to open.

Issue #3432


git-svn-id: http://svn.slicer.org/Slicer4/trunk@22612 3bd1e089-480b-0410-8dfb-8563597acbee