ENH: add node attributes describing the currently active manipulation
Tell which slice viewer is manipulating which fiducial in the
markup node.

Markups.MovingInSliceView holds the LayoutName of the sliceNode associated
with the viewer where the widget is being manipulated.

Markups.MovingSeedNumber holds the seed number being manipulated.

Both are cleared when the seed is not being manipulated by a widget.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@22637 3bd1e089-480b-0410-8dfb-8563597acbee